fix: 회원정보 수정 기능 수정 및 클래스명 변경
