+++ user-service/src/main/java/com/wesell/userservice/service/UserService.java
@@ -1,7 +1,7 @@
 package com.wesell.userservice.service;
 
-import com.wesell.userservice.dto.SignupRequestDto;
-import com.wesell.userservice.dto.responseDto;
+import com.wesell.userservice.dto.request.SignupRequestDto;
+import com.wesell.userservice.dto.response.ResponseDto;
 import com.wesell.userservice.exception.UserNotFoundException;
 import com.wesell.userservice.domain.entity.User;
 import com.wesell.userservice.domain.repository.UserRepository;
@@ -10,7 +10,6 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,18 +19,18 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public responseDto findUser(String uuid) throws UserNotFoundException { // uuid로 유저 조회
+    public ResponseDto findUser(String uuid) throws UserNotFoundException { // uuid로 유저 조회
         Optional<User> userOptional = userRepository.findByOneId(uuid);
         User user = userOptional.orElseThrow(() ->
                 new UserNotFoundException("유저 정보를 찾을 수 없습니다."));
-        return responseDto.of(user);
+        return ResponseDto.of(user);
     }
 
-    public List<responseDto> findUsers() throws UserNotFoundException { // 유저 전체 조회
+    public List<ResponseDto> findUsers() throws UserNotFoundException { // 유저 전체 조회
         Optional<List<User>> usersOptional = userRepository.findAll();
         List<User> userList = usersOptional.orElseThrow(() ->
                 new UserNotFoundException("유저 정보를 찾을 수 없습니다."));
-        return responseDto.of(userList);
+        return ResponseDto.of(userList);
     }
 
     @Transactional
@@ -64,25 +63,21 @@ public class UserService {
         return userRepository.findNicknameByUuid(uuid);
     }
 
-
     @Transactional
     public void updateUser(String uuid, SignupRequestDto signupRequestDTO) {
         Optional<User> optionalUser = userRepository.findByOneId(uuid);
 
         if (optionalUser.isPresent()) {
             User updateduser = optionalUser.get();
-            // 업데이트할 필드를 DTO에서 가져와 업데이트합니다.
-            User user = User.builder()
-                    .id(updateduser.getId())
-                    .name(signupRequestDTO.getName())
-                    .nickname(signupRequestDTO.getNickname())
-                    .phone(signupRequestDTO.getPhone())
-                    .agree(signupRequestDTO.isAgree())
-                    .uuid(signupRequestDTO.getUuid())
-                    .createdAt(LocalDateTime.now())
-                    .build();
-            // 기타 필드 업데이트...
-            // UserRepository를 통해 엔터티를 저장합니다.
+//            User user = User.builder()
+//                    .id(updateduser.getId())
+//                    .name(signupRequestDTO.getName())
+//                    .nickname(signupRequestDTO.getNickname())
+//                    .phone(signupRequestDTO.getPhone())
+//                    .agree(signupRequestDTO.isAgree())
+//                    .createdAt(LocalDateTime.now())
+//                    .build();
+            User user = updateduser.changeName(signupRequestDTO.getName());
             userRepository.update(user);
         } else {
             throw new EntityNotFoundException("User not found with uuid: " + uuid);
